requirements.txt
ansi2html==1.8.0
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
fonttools==4.42.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.4
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.15.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
retrying==1.3.4
schema==0.7.4
six==1.16.0
tenacity==8.2.2
tomli==2.0.1
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
visdcc==0.0.50
Werkzeug==2.2.3
zipp==3.15.0
Django==3.2.21
djangorestframework==3.14.0